feat: successfully run pages:preview
