+++ .gitignore
@@ -27,6 +27,7 @@ yarn-error.log*
 
 # local env files
 .env*.local
+.env.dev
 
 # vercel
 .vercel
